# perf: optimize chat API to eliminate auth duplication and improve DB operations