--- a/lib/actions/chat.ts
+++ b/lib/actions/chat.ts
@@ -149,13 +149,13 @@ export async function createChatWithFirstMessage(
 /**
  * Upsert a message to a chat
  * @param userId - Required but not used for access check (assumes already authorized)
- * 
+ *
  * IMPORTANT: This function assumes the caller has already performed authorization checks.
  * It is only called from:
  * 1. API routes after authentication (app/api/chat/route.ts)
  * 2. Stream handlers after chat ownership verification
  * 3. Internal functions that have already verified access
- * 
+ *
  * DO NOT call this function directly from untrusted contexts.
  */
 export async function upsertMessage(