--- a/lib/streaming/helpers/types.ts
+++ b/lib/streaming/helpers/types.ts
@@ -10,4 +10,5 @@ export interface StreamContext {
   initialChat: (Chat & { messages: UIMessage[] }) | null
   abortSignal?: AbortSignal
   parentTraceId?: string
+  isNewChat?: boolean
 }