--- a/lib/streaming/types.ts
+++ b/lib/streaming/types.ts
@@ -10,4 +10,5 @@ export interface BaseStreamConfig {
   trigger?: 'submit-user-message' | 'regenerate-assistant-message'
   messageId?: string
   abortSignal?: AbortSignal
+  isNewChat?: boolean
 }