… lib/actions/__tests__/chat.test.ts
… lib/streaming/create-chat-stream-response.ts
… lib/streaming/helpers/handle-stream-finish.ts
… lib/streaming/helpers/prepare-messages.ts
… lib/streaming/message-preparation.ts
… lib/utils/perf-logging.ts
… lib/utils/perf-tracking.ts
… scripts/chat-cli.ts