## Summary
- Reduced authentication calls from 5 to 1 per request (~2.4s savings)
- Optimized DB operations with transactions (reduced from 5+ to 2-3 operations)
- Added `isNewChat` flag to skip unnecessary chat loading for new conversations
- Made `userId` required parameter to prevent accidental auth duplication
- Added configurable performance logging via `ENABLE_PERF_LOGGING` env variable

## Changes

### Phase 1: Function Naming
- Renamed functions for clarity (`saveMessage` → `upsertMessage`, `getChat` → `loadChat`)

### Phase 2: Auth Optimization
- Made `userId` a required parameter throughout the codebase
- Removed redundant auth calls in already-authorized contexts
- Single auth call at API route entry point

### Phase 3: New Chat Optimization
- Added `isNewChat` flag from client
- Skip `loadChat` for new chats
- Create chat and first message in single transaction

### Phase 4: DB Operation Optimization
- Added `createChatWithFirstMessageTransaction` for atomic operations
- Removed redundant access checks in `upsertMessage`
- Parallel DB operations where possible

## Performance Improvements

### Before:
- Auth calls: 5 (~2.4s)
- DB operations: 5+
- Total request time: ~15.5s

### After:
- Auth calls: 1 (~0.5s)
- DB operations: 2-3
- Significant reduction in total request time

## Testing
- Added comprehensive test suite for all chat actions (24 test cases)
- All existing tests pass
- Type checking, linting, and formatting verified

## Breaking Changes
None - All changes are backward compatible

🤖 Generated with [Claude Code](https://claude.ai/code)